Oh shit here is some stuffs. I should really start doing this properly.
I have added the basic functionality to the recipe list and detail pages. This includes adding 2 tables to the database - recipe and method (method being an instruction for a recipe).

Additionally, I have added a global-ish function to initiateCooking. This navigates you to the cooking page and loads in the recipe saved using SharedPreferences.